requirements.txt
fastapi
uvicorn[standard]
python-dotenv
agentscope
ollama
requests
python-multipart
pydantic